basic functionality of 24V Industrial I/O Shield
This sketch, intended to provide insight on basic functionality of the shield, sets up the ADC and I/O expander, with Input 0 (B0) set up as an input, and A0 set up as an output.  The output will mirror the input (with a 1 second delay), and will also spit out values from the ADC (all 4 channels).  The values that are non-zero are due to the tiny amount of leakage through the clamping diode, but any input will overcome the leakage.